requirements.txt
Flask
opencv-python
face_recognition
Pillow
